requirements.txt
django-countries==4.4.0
django_phonenumbers==1.0.1
django-pandas==0.6.6
pandas==1.5.2
plotly==5.19.0
scikit-learn==1.4.0
channels==4.0.0
daphne==4.1.0
matplotlib==3.6.2
seaborn==0.13.2
numpy==1.24.1
scipy==1.12.0
joblib==1.3.2
reportlab==3.6.8
argparse==1.4.0